refactor: clearer file names and organization
Three files was named KartleggingssporsmalForm. Renaming files and splitting some code to multiple files.

- Move shared kartleggingssporsmal question and schema definitions into separate modules for schema and questions. Also separate files for field order and conditional field logic.
- Rename form snapshot mapping utility file.
