+++ src/forms/kartleggingssporsmal/conditional-fields-logic.ts
@@ -1,0 +1,12 @@
+import type { KartleggingssporsmalForm } from "./formSchema";
+
+export function shouldIncludeTilbakeTilJobbBegrunnelseField(
+  hvorSannsynligTilbakeTilJobben:
+    | KartleggingssporsmalForm["hvorSannsynligTilbakeTilJobben"]
+    | "",
+): boolean {
+  return (
+    hvorSannsynligTilbakeTilJobben === "1b" ||
+    hvorSannsynligTilbakeTilJobben === "1c"
+  );
+}
