+++ src/forms/kartleggingssporsmal/fieldIdsDisplayOrder.ts
@@ -1,0 +1,8 @@
+import type { KartleggingsspormalFormQuestionId } from "./formQuestions";
+
+export const fieldIdsDisplayOrder: KartleggingsspormalFormQuestionId[] = [
+  "hvorSannsynligTilbakeTilJobben",
+  "hvorSannsynligTilbakeTilJobbenBegrunnelse",
+  "samarbeidOgRelasjonTilArbeidsgiver",
+  "naarTilbakeTilJobben",
+];
