+++ src/forms/kartleggingssporsmal/formQuestions.ts
@@ -1,26 +1,7 @@
-import { type ZodType, z } from "zod/v4";
 import type { RadioGroupQuestion } from "@/components/form-components/RadioGroup";
 import type { TextQuestion } from "@/components/form-components/TextArea";
 
-export const fieldIdsDisplayOrder: KartleggingsspormalFormQuestionId[] = [
-  "hvorSannsynligTilbakeTilJobben",
-  "hvorSannsynligTilbakeTilJobbenBegrunnelse",
-  "samarbeidOgRelasjonTilArbeidsgiver",
-  "naarTilbakeTilJobben",
-];
-
-export function shouldIncludeTilbakeTilJobbBegrunnelseField(
-  hvorSannsynligTilbakeTilJobben:
-    | KartleggingssporsmalForm["hvorSannsynligTilbakeTilJobben"]
-    | "",
-): boolean {
-  return (
-    hvorSannsynligTilbakeTilJobben === "1b" ||
-    hvorSannsynligTilbakeTilJobben === "1c"
-  );
-}
-
-const radioGroupQuestions = {
+export const radioGroupQuestions = {
   hvorSannsynligTilbakeTilJobben: {
     type: "RADIO_GROUP",
     label:
@@ -67,31 +48,5 @@ export const kartleggingssporsmalFormQuestions = {
   ...textQuestions,
 } as const;
 
-type KartleggingsspormalFormQuestionId =
+export type KartleggingsspormalFormQuestionId =
   keyof typeof kartleggingssporsmalFormQuestions;
-
-function getRadioGroupOptionIds(
-  radioFieldId: keyof typeof radioGroupQuestions,
-) {
-  return radioGroupQuestions[radioFieldId].options.map((option) => option.id);
-}
-
-export const kartleggingssporsmalFormSchema = z.object({
-  hvorSannsynligTilbakeTilJobben: z.enum(
-    getRadioGroupOptionIds("hvorSannsynligTilbakeTilJobben"),
-    "Feltet er påkrevd",
-  ),
-  hvorSannsynligTilbakeTilJobbenBegrunnelse: z.string(),
-  samarbeidOgRelasjonTilArbeidsgiver: z.enum(
-    getRadioGroupOptionIds("samarbeidOgRelasjonTilArbeidsgiver"),
-    "Feltet er påkrevd",
-  ),
-  naarTilbakeTilJobben: z.enum(
-    getRadioGroupOptionIds("naarTilbakeTilJobben"),
-    "Feltet er påkrevd",
-  ),
-} satisfies Record<KartleggingsspormalFormQuestionId, ZodType>);
-
-export type KartleggingssporsmalForm = z.infer<
-  typeof kartleggingssporsmalFormSchema
->;
